Add NVFP emulation components and update backend functions
- Introduced new `w3_w4.py` script for searching optimal W3 and W4 values in NVFP emulation.
- Added `emulation` package with core classes and utilities for NVFP4 emulation, including `SearchNVFPEmulationState` and `MMAEngine`.
- Implemented hardware configuration management in `config.py` for NVFP4 emulation.
- Updated `nvfp.py` to include `w_reduce` parameter in the NVFPQuantState and adjusted function calls accordingly.
- Enhanced rounding strategies and utility functions for NVFP4 operations.

+++ search/w_to_1.py
@@ -22,7 +22,7 @@ ab_dtype = torch.float16
 A = torch.zeros(256, 256, dtype=ab_dtype, device="cuda")
 B = torch.zeros(256, 256, dtype=ab_dtype, device="cuda")
 
-quant_fn, real_fn, emul_fn, meta = build_gemm_compare_fns(args.backend)
+quant_fn, real_fn, emul_fn, meta = build_gemm_compare_fns(args.backend, out_dtype=ab_dtype)
 
 
 def set_group(X: torch.Tensor, i: int, j: float, group_size: int):
